+++ src/com/intellij/idea/plugin/hybris/project/HybrisProjectImportProvider.java
@@ -22,11 +22,7 @@ import com.intellij.ide.util.projectWizard.ModuleWizardStep;
 import com.intellij.ide.util.projectWizard.ProjectWizardStepFactory;
 import com.intellij.ide.util.projectWizard.WizardContext;
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils;
-import com.intellij.idea.plugin.hybris.project.wizard.CheckRequiredPluginsStep;
-import com.intellij.idea.plugin.hybris.project.wizard.HybrisWorkspaceRootStep;
-import com.intellij.idea.plugin.hybris.project.wizard.InformationStep;
-import com.intellij.idea.plugin.hybris.project.wizard.SelectHybrisModulesToImportStep;
-import com.intellij.idea.plugin.hybris.project.wizard.SelectOtherModulesToImportStep;
+import com.intellij.idea.plugin.hybris.project.wizard.*;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.projectImport.ProjectImportBuilder;
 import com.intellij.projectImport.ProjectImportProvider;
@@ -48,7 +44,7 @@ public class HybrisProjectImportProvider extends ProjectImportProvider {
         return new ModuleWizardStep[]{
             new CheckRequiredPluginsStep(context),
             new InformationStep(context),
-            new HybrisWorkspaceRootStep(context),
+            new ProjectImportWizardRootStep(context),
             new SelectHybrisModulesToImportStep(context),
             new SelectOtherModulesToImportStep(context),
             stepFactory.createProjectJdkStep(context)